refactor(core): move json reporter
